==> 01_loading_and_normalizingPulmNETcaseseries.Rmd <==
---
title: "01-loading_and_normalizingPulmNETcaseseries"
author: "Wenzel Hackeng"
date: "29/10/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## **Pre-processing**
First call the required packages
```{r message=FALSE, warning=FALSE}
library("ChAMP")
library("readr")
library("plyr")
```

```{r eval=FALSE}
dir.create("RAW")
dir.create("./RAW/EGAD00010001720")
```

# Get credentials through EGA and authors Alcala et al. 2019, run in terminal
pyega3 -cf /CREDENTIALS_FILE.json fetch EGAD00010001720 --saveto /RAW/EGAD00010001720
Collect all histopathological diagnosed carcinoid idat files. 

# ChAMP load data 
During filtering, manually check proportion of failed probes to see if there are bad samples.
```{r message=FALSE, warning=FALSE}
dir.create("01_preprocessing")
filesEGAD00010001720<-'./RAW/EGAD00010001720/'
set.seed(11)
myloadEGAD00010001720 <- champ.load(directory = filesEGAD00010001720,
                              method="ChAMP",
                              methValue="B",
                              autoimpute=FALSE, # not default
                              filterDetP=TRUE,
                              ProbeCutoff=0,
                              SampleCutoff=0.1,
                              detPcut=0.01,
                              filterBeads=TRUE,
                              beadCutoff=0.05,
                              filterNoCG=TRUE,
                              filterSNPs=TRUE,
                              population=NULL,
                              filterMultiHit=TRUE,
                              filterXY=TRUE,
                              force=FALSE,
                              arraytype="EPIC")
```
# ChAMP Quality Control 
Check density plots for sample quality. 
```{r message=FALSE, warning=FALSE}

# 1. Quality Control files
dir.create("./01_preprocessing/CHAMP_QCimages")
champ.QC(beta = myloadEGAD00010001720$beta,
         pheno=myloadEGAD00010001720$pd$Sample_Group,
         mdsPlot=TRUE,
         densityPlot=TRUE,
         dendrogram=TRUE,
         PDFplot=TRUE,
         Rplot=TRUE,
         Feature.sel="None",
         resultsDir = "./01_preprocessing/CHAMP_QCimages/EGAD00010001720/")
```

# ChAMP Normalisation
Normalisation, BMIQ is used voor data as Karphathakis' data was BMIQ normalized (BMIQ best method)

```{r message=FALSE, warning=FALSE}
dir.create("./01_preprocessing/Normalisation")
mynormEGAD00010001720<-champ.norm(beta=myloadEGAD00010001720$beta, rgSet=myloadEGAD00010001720$rgSet, mset=myloadEGAD00010001720$mset, 
                    resultsDir = "./01_preprocessing/Normalisation/EGAD00010001720/", 
                    method="BMIQ", 
                    plotBMIQ=TRUE, 
                    arraytype="EPIC",
                    cores=4)
```

# Randomly select samples in case of technical replicates
```{r message=FALSE, warning=FALSE}
PD_EGAD00010001720 <- read_csv("./PD_FILES/PD_EGAD00010001720.csv", 
    skip = 7)

# select one of two duplicates
set.seed(11)
PD_EGAD00010001720_UNIQ <- ddply(PD_EGAD00010001720,.(Name_True),
    function(x) {
          x[sample(nrow(x),size=1),]
    })

# select betas after filtering for unique tumors
selectedcases<-PD_EGAD00010001720_UNIQ$Sample_Name
mynormEGAD00010001720_UNIQ<-as.data.frame(mynormEGAD00010001720[,selectedcases])

# these raw case numbers were selected for the test cohort
as.matrix(PD_EGAD00010001720_UNIQ[,c("Sentrix_ID", "Sentrix_Position")])
```

# Save normalized data for later use
```{r message=FALSE, warning=FALSE}
dir.create("./01_preprocessing/PreProcessedDataSets")
dir.create("./01_preprocessing/PreProcessedDataSets/EGAD00010001720")
save(myloadEGAD00010001720, mynormEGAD00010001720_UNIQ, PD_EGAD00010001720_UNIQ, file = './01_preprocessing/PreProcessedDataSets/EGAD00010001720/preprocessedEGAD00010001720.rData')

```
